backend/requirements.txt
# Core FastAPI framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
asyncpg>=0.29.0
databases>=0.8.0

# Environment and configuration
python-dotenv>=1.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# External services
supabase>=2.0.0
requests>=2.31.0

# Data validation and serialization
pydantic>=2.0.0

# Machine Learning
joblib>=1.3.0
scikit-learn>=1.3.0

# Optional: PostgreSQL adapter (can be problematic on some systems)
# psycopg2-binary>=2.9.0

# Development and testing (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0